[OC-379] Detail screen with click tracker url triggered

+++ MediaCorpPOC/app/src/main/java/com/t/mediacorp2359pocs/model/oc379/response/ItemResponse.kt
@@ -15,7 +15,7 @@ data class ItemResponse(
     val rankingAlgorithm: String = "",
 
     @SerializedName("ranking_score")
-    val rankingScore: Int = 0,
+    val rankingScore: Float = 0F,
 
     @SerializedName("content_id")
     val contentId: String = "",
